requirements.txt
alembic==1.4.3
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.1.7
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
flask-restx==0.3.0
importlib-metadata==2.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.24
Werkzeug==1.0.1
zipp==1.2.0
yfinance==0.1.59
redis==3.5.3
psycopg2-binary==2.8.6
psycopg2==2.8.6
flask-cors==3.0.10
scipy==1.4.1
numpy==1.18.5
pandas==0.24.2